# api-docs/swagger.yaml
basePath: /
definitions:
...
  admin.emailLoginSettingsResponse:
    properties:
      enabled:
        type: boolean
      from_email:
        example: mailer@example.com
        type: string
      from_name:
        example: 即应
        type: string
      smtp_host:
        example: smtp.example.com
        type: string
      smtp_password:
        type: string
      smtp_password_configured:
        type: boolean
      smtp_port:
        example: 587
        type: integer
      smtp_security:
        example: starttls
        type: string
      smtp_username:
        example: mailer@example.com
        type: string
    type: object
  admin.errorBody:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  admin.errorEnvelope:
    properties:
      error:
        $ref: '#/definitions/admin.errorBody'
      success:
        example: false
        type: boolean
    type: object
...
  admin.successEnvelope:
    properties:
      data: {}
      success:
        example: true
        type: boolean
    type: object
  admin.testEmailLoginSettingsRequest:
    properties:
      recipient_email:
        example: admin@example.com
        type: string
    type: object
...
  admin.updateEmailLoginSettingsRequest:
    properties:
      enabled:
        type: boolean
      from_email:
        type: string
      from_name:
        type: string
      smtp_host:
        type: string
      smtp_password:
        type: string
      smtp_port:
        type: integer
      smtp_security:
        type: string
      smtp_username:
        type: string
    type: object
...
info:
  contact: {}
  description: 私有部署企业协作产品的服务端 API。当前阶段包含管理员登录、管理员创建普通用户、普通用户登录。
  title: AI 原生企业协作服务 API
  version: 0.1.0
paths:
...
  /api/admin/settings/email-login:
    get:
      description: 管理员读取邮箱验证码登录和完整 SMTP 设置，包括已保存的 SMTP 密码。
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/admin.successEnvelope'
            - properties:
                data:
                  $ref: '#/definitions/admin.emailLoginSettingsResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/admin.errorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/admin.errorEnvelope'
      summary: 获取邮箱验证码登录设置
      tags:
      - 管理员设置
    put:
      consumes:
      - application/json
      description: 管理员配置邮箱验证码登录和 SMTP。省略 smtp_password 时保留原密码。
      parameters:
      - description: 邮箱登录设置
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/admin.updateEmailLoginSettingsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/admin.successEnvelope'
            - properties:
                data:
                  $ref: '#/definitions/admin.emailLoginSettingsResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/admin.errorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/admin.errorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/admin.errorEnvelope'
      summary: 更新邮箱验证码登录设置
      tags:
      - 管理员设置
  /api/admin/settings/email-login/test:
    post:
      consumes:
      - application/json
      description: 管理员使用已保存的 SMTP 配置发送测试邮件。失败时返回可诊断的错误原因。
      parameters:
      - description: 测试收件邮箱
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/admin.testEmailLoginSettingsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/admin.successEnvelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/admin.errorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/admin.errorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/admin.errorEnvelope'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/admin.errorEnvelope'
      summary: 发送 SMTP 测试邮件
      tags:
      - 管理员设置
...
securityDefinitions:
  UserSession:
    description: 使用 user_session=<token> 格式的会话 Cookie。
    in: header
    name: Cookie
    type: apiKey
swagger: "2.0"